Comments adjusted
